+++ src/main/java/com/review/shop/dto/orders/OrderItemDetailDTO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class OrderItemDetailDTO {
-    private long order_item_id;
-    private long product_id;
+    private int order_item_id;
+    private int product_id;
     private String product_name;
     private int quantity;
     private int product_price;
